+++ src/index.ts
@@ -1,10 +1,9 @@
 import 'reflect-metadata';
 
 export * from './thresh/thresh';
-export * from './methods/route';
-export * from './methods/method';
-export * from './methods/description';
+export * from './methods';
 export * from './enum';
+export * from './services';
 
 // Types
 import { Request as Req, Response as Res, NextFunction as Next } from 'express';